Fix edge-case error when adding annotation to drawing
This fixes a `ReferenceError: StructRNA of type Object has been removed`
error that prevented adding new annotations, particularly dimension
lines, to a drawing.  It appears that, in some cases, the call to
`drawing_tool.reload_representation` can cause the original `obj` to
become invalid, which then results in a crash in `collector.assign`
when accessing the `obj.users_collection` property.  To prevent this
error, the `obj` variable needs to be updated with the current version
of the object after calling `reload_representation`.